chaotic-encryption-app/backend/requirements.txt
Flask
Flask-CORS
opencv-python
numpy
Pillow
scipy
Werkzeug
gunicorn
pytest
hypothesis
pycryptodome
